update procfile, add static page, update worker
