+++ Procfile
@@ -1,1 +1,2 @@
-worker: npm start
+web: npm start
+worker: npm start oracle
