+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node -r esm index.js",
+    "oracle": "node -r esm oracle.js",
     "query": "node -r esm query.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
